Recreates the Workspace fixing path issues.
There was a problem that was nos addressed by an issue related to path when creating from setup.sh. Now it adds the new Makefile so that it keeps things more organized and it also reduces the number of files at the root of the backend part of the project.