Test routines working
Until the moment there is only the test to verify if a
exercise is created correctly. To isolate the tests
from the REST api a clone script from the app is generated
with a localstorage adapter instead of the RESTAdapter.
Then this other version is added for the qunit test.